feat(create): preview toolbar progressive blur, themed by preset
Extract the header's layered blur stack into a shared ProgressiveBlur
component and use it for the create preview toolbar in place of the flat
background fade. The toolbar's tint resolves the previewed design
system's --color-bg in the preview's own light/dark mode via a scoped
DesignSystemProvider, so the fade matches the iframe background instead
of the site theme.
