+++ www/src/styles.css
@@ -61,10 +61,11 @@
    technique as shadcn's scroll-fade utilities (scroll-driven animation of a
    registered @property), retargeted from a mask to backdrop-filter. The scroll timeline ramps
    --blur-progress 0 → 1 across the first header-height of scroll; the blur layers
-   scale their radius by it and the tint scales its opacity (see header.tsx), so the
-   container never carries a fractional opacity — which would make it an opacity
-   group and break backdrop sampling. Where scroll timelines are unsupported the blur
-   stays at 0 and the header falls back to a solid background (see header-blur-fallback). */
+   scale their radius by it and the tint scales its opacity (see ProgressiveBlur in
+   components/progressive-blur.tsx), so the container never carries a fractional
+   opacity — which would make it an opacity group and break backdrop sampling. Where
+   scroll timelines are unsupported the blur stays at 0 and the header falls back to
+   a solid background (see header-blur-fallback). */
 @property --blur-progress {
   syntax: '<number>';
   inherits: true;
@@ -99,11 +100,11 @@
   }
 }
 
-/* header-blur-dither — a low-opacity tiled noise that dithers the navbar tint's
-   near-black gradient so 8-bit color banding doesn't show as horizontal steps.
-   Opacity is ramped by --blur-progress inline (see header.tsx); masked to the upper
-   band where the tint sits. */
-@utility header-blur-dither {
+/* progressive-blur-dither — a low-opacity tiled noise that dithers the blur stack's
+   near-black tint gradient so 8-bit color banding doesn't show as horizontal steps.
+   Opacity is ramped by --blur-progress inline (see components/progressive-blur.tsx);
+   masked to the upper band where the tint sits. */
+@utility progressive-blur-dither {
   background-image: url("data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20width='140'%20height='140'%3E%3Cfilter%20id='n'%3E%3CfeTurbulence%20type='fractalNoise'%20baseFrequency='0.85'%20numOctaves='2'%20stitchTiles='stitch'/%3E%3CfeColorMatrix%20type='saturate'%20values='0'/%3E%3C/filter%3E%3Crect%20width='100%25'%20height='100%25'%20filter='url(%23n)'/%3E%3C/svg%3E");
   background-size: 140px 140px;
   mask-image: linear-gradient(to top, transparent 10%, #000 60%);
